requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.1
beautifulsoup4==4.12.2
black==23.1.0
bs4==0.0.1
cachetools==5.3.0
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
compress==0.0.3
constantly==15.1.0
country-list==1.0.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
crispy-tailwind==0.5.0
cryptography==39.0.1
cssbeautifier==1.14.7
dalle43==0.0.0.6
daphne==4.0.0
defusedxml==0.7.1
distro==1.8.0
dj-pagination==2.5.0
dj-places==5.2.1
Django==4.1.7
django-admin-tailwind==0.0.23
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap-modal-forms==2.2.1
django-browser-reload==1.7.0
django-compressor==4.3.1
django-countable-field==1.3
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==23.1
django-galleryfield==2.1.3
django-geojson==4.0.0
django-htmx==1.12.2
django-leaflet==0.28.3
django-libsass==0.9
django-location-field==2.7.0
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-private-chat2==1.0.2
django-render-block==0.9.2
django-resized==1.0.2
django-sendfile2==0.7.0
django-starfield==1.0.post2
django-widget-tweaks==1.4.12
django3-livesync==1
django_debug_toolbar==3.8.1
djlint==1.19.16
EditorConfig==0.12.3
exceptiongroup==1.1.1
GDAL @ file:///Users/runner/miniforge3/conda-bld/gdal-split_1678201918264/work/build/swig/python
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsbeautifier==1.14.7
jsonfield==3.1.0
kolo==2.4.4
libsass==0.22.0
Markdown==3.4.3
msgpack==1.0.5
mypy-extensions==1.0.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1675642719423/work
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
phonenumberslite==8.13.6
Pillow==9.4.0
pinax-messages==3.0.0
platformdirs==3.0.0
protobuf==4.22.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
redis==4.5.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==12.6.0
rjsmin==1.2.1
rsa==4.9
selenium==4.8.2
service-identity==21.1.0
shell_gpt==0.9.0
shortuuid==1.0.11
simple-django-htmx==1.0.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
txaio==23.1.1
typer==0.7.0
typing_extensions==4.5.0
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==1.26.14
watchdog==3.0.0
webdriver-manager==3.8.5
websocket-client==1.5.1
wsproto==1.2.0
zope.interface==6.0
